Refactoring, decomposition
***
frontend
