+++ frontend/app/lib/components/board/graph/graphBoardClientWrapper.tsx
@@ -4,6 +4,7 @@ import { useEffect } from 'react';
 import settingsStore from '@/app/stores/settingsStore';
 import { ReactFlowProvider } from '@xyflow/react';
 import GraphBoard from '@/app/lib/components/board/graph/graphBoard';
+import { GraphBoardProvider } from '@/app/lib/components/board/graph/graphBoardContext';
 
 export default function GraphBoardClientWrapper() {
   useEffect(() => {
@@ -12,7 +13,9 @@ export default function GraphBoardClientWrapper() {
 
   return (
     <ReactFlowProvider>
-      <GraphBoard />
+      <GraphBoardProvider>
+        <GraphBoard />
+      </GraphBoardProvider>
     </ReactFlowProvider>
   );
 }
