+++ frontend/app/lib/components/board/simple/tools/baseTools/pencilTool.tsx
@@ -10,7 +10,7 @@ import {
 import settingsStore from '@/app/stores/settingsStore';
 import FTooltip from '@/app/lib/components/foggyOverrides/fTooltip';
 import { useBoardContext } from '@/app/lib/components/board/simple/boardContext';
-import useTool from '@/app/lib/hooks/useTool';
+import useTool from '@/app/lib/hooks/simpleBoard/useTool';
 import debounce from 'lodash/debounce';
 import useAdaptiveParams from '@/app/lib/hooks/useAdaptiveParams';
