+++ frontend/app/lib/components/board/simple/tools/baseTools/rectTool.tsx
@@ -9,7 +9,7 @@ import {
 import settingsStore from '@/app/stores/settingsStore';
 import FTooltip from '@/app/lib/components/foggyOverrides/fTooltip';
 import { useBoardContext } from '@/app/lib/components/board/simple/boardContext';
-import useTool from '@/app/lib/hooks/useTool';
+import useTool from '@/app/lib/hooks/simpleBoard/useTool';
 import useAdaptiveParams from '@/app/lib/hooks/useAdaptiveParams';
 
 export default function RectTool() {
