+++ frontend/app/lib/hooks/simpleBoard/useStageSize.ts
